requirements_venv.txt
numpy
pygame
torchrl
matplotlib